// source/store/settings/types.ts
export type SettingsStore = {
    equalizer: boolean;
    equalizerSurround: boolean;
    equalizerEffects: boolean;
    equalizerAnalyser: boolean;
    equalizerCompressor: boolean;
    equalizerCompressorThreshold: number;
    equalizerCompressorKnee: number;
    equalizerCompressorRatio: number;
    equalizerCompressorAttack: number;
    equalizerCompressorRelease: number;
    scrobbler: boolean;
    subscribeToGroup: boolean;
};